refactored reduce shape to dset_lib
